refactor(output): extract attachment tests to sibling file
